Refactor numpy import checks
