+++ traits/tests/test_array.py
@@ -13,14 +13,10 @@ from __future__ import absolute_import
 
 import unittest
 
-try:
-    import numpy
-except ImportError:
-    numpy_available = False
-else:
-    numpy_available = True
-
 from traits.api import Array, Bool, HasTraits
+from traits.util.import_handler import import_handler
+
+numpy, numpy_available = import_handler('numpy')
 
 
 if numpy_available:
