+++ traits/tests/test_integer.py
@@ -24,14 +24,10 @@ import unittest
 
 import six
 
-try:
-    import numpy
-except ImportError:
-    numpy_available = False
-else:
-    numpy_available = True
-
 from traits.api import HasTraits, Int, TraitError
+from traits.util.import_handler import import_handler
+
+numpy, numpy_available = import_handler('numpy')
 
 if six.PY2:
     LONG_TYPE = long
